requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
pathlib2>=2.3.7

# Data processing
pandas>=2.0.0
pyyaml>=6.0
pillow>=10.0.0

# Progress bars
tqdm>=4.65.0

# YOLO and training
ultralytics>=8.0.0

# Image Quality Assessment (for image_quality_grader.py)
opencv-contrib-python>=4.8.0  # For BRISQUE/NIQE from cv2.quality
torch>=2.0.0                   # For deep learning metrics
torchvision>=0.15.0           # For image transforms
pyiqa>=0.1.7                  # Comprehensive IQA library (NIQE, BRISQUE, PIQE, MUSIQ, CONTRIQUE, CLIPIQA)

# Optional: for faster processing
# timm>=0.9.0                 # Included with pyiqa
# clip @ git+https://github.com/openai/CLIP.git  # Included with pyiqa
